requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose
python-multipart
pydantic>=2.5,<3
pydantic-settings>=2.2,<3
